requirements.txt
faiss-cpu==1.11.0
fastapi[standard]==0.115.12
ollama==0.4.8
openai==1.81.0
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==4.1.0
streamlit==1.45.1
torch==2.7.0
